+++ HabboHotel/Rooms/RoomManager.cs
@@ -29,6 +29,7 @@ public class RoomManager : IRoomManager
     private readonly IGameClientManager _clientManager;
     private readonly IItemLoader _itemLoader;
     private readonly IRoomItemPersistenceService _roomItemPersistenceService;
+    private readonly IRoomItemPlacementValidatorService _roomItemPlacementValidatorService;
     private readonly IChatManager _chatManager;
     private readonly IBotManager _botManager;
     private readonly IRoomService _roomService;
@@ -51,6 +52,7 @@ public class RoomManager : IRoomManager
         IRoomFactory roomFactory,
         IItemLoader itemLoader,
         IRoomItemPersistenceService roomItemPersistenceService,
+        IRoomItemPlacementValidatorService roomItemPlacementValidatorService,
         IGameClientManager gameClientManager,
         IGroupManager groupManager,
         IRoomService roomService,
@@ -71,6 +73,7 @@ public class RoomManager : IRoomManager
         _roomFactory = roomFactory;
         _itemLoader = itemLoader;
         _roomItemPersistenceService = roomItemPersistenceService;
+        _roomItemPlacementValidatorService = roomItemPlacementValidatorService;
         _clientManager = gameClientManager;
         _groupManager = groupManager;
         _roomService = roomService;
@@ -212,7 +215,7 @@ public class RoomManager : IRoomManager
 
     private Room CreateRoomInstance(RoomData data)
     {
-        return new Room(data, _clientManager, _database, _itemLoader, _roomItemPersistenceService, _groupManager, _roomService, _chatManager, _botManager, _achievementService, _questService, _cacheManager, _languageManager, _itemTeleporterFinder, _itemHopperFinder, _badgeManager, _userDataFactory, this, _loggerFactory);
+        return new Room(data, _clientManager, _database, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _groupManager, _roomService, _chatManager, _botManager, _achievementService, _questService, _cacheManager, _languageManager, _itemTeleporterFinder, _itemHopperFinder, _badgeManager, _userDataFactory, this, _loggerFactory);
     }
 
     private static void DisposeRoom(Room room) => room.Dispose();